+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to the
 
 - add `Data.String.Unindent.unindent` function: runtime unindent string values
 - add `Data.String.Indoc.indoc` quasi-quoter: compile time unindent string literals
+- add `Data.String.Format` and parser: Rust-style `format_args` format strings